+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "gtest"]
 	path = gtest
-	url = git@github.com:google/googletest.git
+	url = https://github.com/google/googletest.git
