+++ environment.yml
@@ -6,65 +6,62 @@ channels:
 
 dependencies:
   - armadillo==8.200.0
-  - blas==1.1=openblas
-  - bullet==2.86.1=0
+  - blas==1.1
+  - bullet==2.86.1
   - bz2file==0.98
-  - bzip2==1.0.6=1
+  - bzip2==1.0.6
   - cmake>=3.10
-  - cspice==66=h470a237_3
-  - curl==7.60.0=0
-  - doxygen==1.8.14=0
-  - eigen==3.3.3=0
-  - embree==2.14.0=0
+  - cspice==66
+  - curl==7.60.0
+  - doxygen==1.8.14
+  - eigen==3.3.3
+  - embree==2.14.0
   - geos==3.5.1
-  - geotiff==1.4.2=1
+  - geotiff==1.4.2
   - gmm==5.0
-  - gmp==6.1.2=0
-  - gsl==2.2.1=blas_openblas_3
-  - hdf5==1.8.18=2
-  - icu==58.2=0
+  - gmp==6.1.2
+  - gsl==2.2.1
+  - hdf5==1.8.18
+  - icu==58.2
   - jama==125
-  - jpeg==9b=2
+  - jpeg==9b
   - kakadu==1
-  - krb5==1.14.2=0
+  - krb5==1.14.2
   - libpng>=1.6.34
   - libprotobuf==3.5.2
-  - libtiff==4.0.9=0
-  - libxml2==2.9.7=0
+  - libtiff>=4.0.9
+  - libxml2==2.9.7
   - make
-  - mesalib==17.2.0=0
+  - mesalib==17.2.0
   - mysql==5.7.20
-  - mysql-connector-c==6.1.6=0
+  - mysql-connector-c==6.1.6
   - nanoflann==1.2.2
-  - ninja==1.7.2=0
+  - ninja==1.7.2
   - conda-forge/label/gcc7::nn
-  # - numpy==1.13.3=py36_blas_openblas_200
-  - openblas==0.2.19=2
+  - openblas==0.2.19
   - opencv
-  - openssl==1.0.2n=0
+  - openssl==1.0.2n
   - patchelf==0.9
   - pcl==1.8.1
   - pip==9.0.1
   - protobuf==3.5.2
-  # - pyqt==5.6.0
   - python==3.6
   - qhull==7.2.0=0
   - qt=5.9.6
   - qwt=6.1.3
   - setuptools=38.5.1
   - sip==4.18
-  - sqlite==3.13.0=1
-  - suitesparse==4.5.4=blas_openblas_200
-  - superlu==5.2.1=blas_openblas_201
+  - sqlite==3.13.0
+  - suitesparse==4.5.4
+  - superlu==5.2.1
   - tnt==126=0
   - wheel==0.30.0
   - x264==20131218
   - xalan-c==1.11
-  - xerces-c==3.1.4=0
-  - xorg-kbproto==1.0.7=1
+  - xerces-c==3.1.4
+  - xorg-kbproto==1.0.7
   - xorg-libice
   - xorg-libsm
-  - xorg-libx11==1.6.4=6
+  - xorg-libx11==1.6.4
   - xorg-libxi
-  - zlib==1.2.11=0
-
+  - zlib==1.2.11
